Initial Commit
Working sketch that gets the brightness from the device's output.

The way to extract the value was derived via trial and error. To figure
out what 'column' (or seeming column) correlated to the brightness, I
checked what value went up as I shined a light into the sensor and went
down when I covered the sensor.